feat(entity): 更新多个实体类，继承BaseEntity并添加构造器
- 在多个实体类中（如ClusterAlertGroupMapEntity、ClusterAlertHistoryEntity等）继承了BaseEntity类，以统一实体的基础属性和行为。
- 添加了无参构造器、全参构造器和SuperBuilder注解，增强了对象创建的灵活性。
- 更新了时间字段类型为LocalDateTime，以确保时间处理的一致性和准确性。

这些改动提升了实体类的可用性和一致性，确保了整体设计的健壮性。

+++ datasophon-api/src/main/java/com/datasophon/api/controller/v1/cluster/ClusterAlertGroupMapController.java
@@ -67,7 +67,7 @@ public class ClusterAlertGroupMapController {
      * 信息
      */
     @RequestMapping("/info/{id}")
-    public Result<ClusterAlertGroupMapVO> info(@PathVariable("id") Integer id) {
+    public Result<ClusterAlertGroupMapVO> info(@PathVariable("id") Long id) {
         // 调用Service层方法，获取DTO
         ClusterAlertGroupMapDTO dto = clusterAlertGroupMapService.getByIdAsDto(id);
         // Controller层：DTO → VO转换
